+++ src/Services/Catalog/CatalogService.Application/UseCases/Category/Commands/Delete/DeleteCategoryCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using CatalogService.Domain.Interfaces;
-using CatalogService.Infostructure;
+﻿using CatalogService.Application.Exceptions;
+using CatalogService.Application.Interfaces.Repositories;
+using CatalogService.Application.Interfaces.UnitOfWork;
+using CatalogService.Domain.Entities;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -18,6 +20,12 @@ namespace CatalogService.Application.UseCases
 
         public async Task Handle(DeleteCategoryCommand request, CancellationToken token)
         {
+            var existingCategory = await _categoryRepository.AnyAsync(request.Id, token);
+            if (!existingCategory)
+            {
+                throw new NotFoundException(nameof(Category));
+            }
+
             await _categoryRepository.DeleteAsync(request.Id, token);
             await _unitOfWork.SaveChangesAsync();
         }
